requirements.txt
streamlit==1.25.0
python-dotenv==1.0.0
langchain-community==0.1.0
langchain-groq==0.1.0
langchain-core==0.1.0
chromadb==0.4.8
re==2.2.1
uuid==1.30
